refactor(step3): move directory
