+++ src/main/kotlin/racingcar/component/RacingcarGameManager.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.component
 
 import racingcar.model.Car
 import racingcar.model.RacingcarGameInput
